MicroServices/adminOperations/api/requirements.txt
flask ~= 1.1
grpcio-tools ~= 1.30
Jinja2 ~= 2.11
google ~= 3.0.0
protobuf ~= 3.8.0
pytest ~= 5.4